se agrego carpeta .idea a .gitignore (se borro archivo trait_spec.rb y se renombro trait2_spec.rb por trait_spec.rb)
